AX: Move AccessibilityRole into its own header to fix a non-unified build issue in AXTableHelpers.cpp https://bugs.webkit.org/show_bug.cgi?id=297859 rdar://159107755
Reviewed by Joshua Hoffman.

This commit allows use of AccessibilityRole without importing the entirety of AXCoreObject.h.

Renames accessibilityRoleToString to roleToString.

* Source/WebCore/Headers.cmake:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/accessibility/AXCoreObject.h:
(WebCore::accessibilityRoleToString): Deleted.
* Source/WebCore/accessibility/AXLogger.cpp:
(WebCore::operator<<):
* Source/WebCore/accessibility/AXTableHelpers.h:
* Source/WebCore/accessibility/AXTextMarker.cpp:
(WebCore::AXTextMarker::debugDescription const):
* Source/WebCore/accessibility/AXUtilities.cpp:
(WebCore::roleToString):
* Source/WebCore/accessibility/AXUtilities.h:
* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::debugDescriptionInternal const):
* Source/WebCore/accessibility/AccessibilityRole.h: Added.
* Source/WebCore/accessibility/ios/WebAccessibilityObjectWrapperIOS.mm:
(-[WebAccessibilityObjectWrapper _accessibilityWebRoleAsString]):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::debugDescriptionInternal const):

Canonical link: https://commits.webkit.org/299307@main
