+++ Source/WebCore/accessibility/AXUtilities.h
@@ -32,6 +32,7 @@
 namespace WebCore {
 
 enum class AXNotification : uint8_t;
+enum class AccessibilityRole : uint8_t;
 enum class NodeName : uint16_t;
 class ContainerNode;
 class Document;
@@ -74,5 +75,7 @@ const RenderStyle* safeStyleFrom(Element&);
 WTF::TextStream& operator<<(WTF::TextStream&, AXNotification);
 
 void dumpAccessibilityTreeToStderr(Document&);
+
+String roleToString(AccessibilityRole);
 
 } // WebCore
