+++ Source/WebCore/accessibility/AccessibilityObject.cpp
@@ -158,7 +158,7 @@ String AccessibilityObject::debugDescriptionInternal(bool verbose, std::optional
 {
     StringBuilder result;
     result.append("{"_s);
-    result.append("role: "_s, accessibilityRoleToString(role()));
+    result.append("role: "_s, roleToString(role()));
     result.append(", ID "_s, objectID().loggingString());
 
     if (debugOptions) {
